Fix an issue on past releases with beta start dates.
On past releases, we have a table that shows the milestones of past releases.
The beta start date is defined as the date for beta 1, but in the past we started
betas with beta 3, our first 2 betas were devedition only.

As a consequence, the date was null (was falling back to today in the template).
The fix is to use the devedition 1 date if there was no beta 1.