Fix Dockerfile for internationalization
The alpine image no longer installs a real ICU library but a shim.
Timezone data (tzdata package) is also not installed by default.

This caused a DateTime() exception in all the templates where we display dates and times

This patch forces the installation of the real ICU library and the tzdata package

References:
https://hoa.ro/blog/2020-12-08-draft-docker-time-timezone/#set-the-time-zone-of-the-containers
https://wiki.alpinelinux.org/wiki/Release_Notes_for_Alpine_3.16.0#ICU_data_split