Bug 1831355: Incorporate iOS release dates into whattrainitisnow
- Added API endpoints
  - api/release/schedule/ios/?version=release
  - api/release/schedule/ios/?version=beta
- Updated About page
- Added an iOS tab on release pages > 144
- Added IOS schedule class, extends the Release class + tests
- Integrate EOY adjustments

Note to future self:

Given that we don't store any iOS release data in product details,
when you make adjustments to a future schedule for iOS,
like cancelling a dot release, you need to also hardcode these changes in getPastSchedule()

The templating work is also impacted, specifically the github links to commits that also requires fixups when you cancel a release.
Potential TODO: Transmit to the past_release template a variable holding the last dot release of the previous cycle to build the changelog link.